Add small fix in regex; Organize tests
